pythonProject/requirements.txt
APScheduler==3.6.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
idna==2.10
mysqlclient==2.0.3
numpy==1.19.5
pandas==1.2.0
paramiko==2.7.2
pycparser==2.20
PyMySQL==0.10.1
PyNaCl==1.4.0
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
selenium==3.141.0
sentry-sdk==0.19.5
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
tornado==6.1
tzlocal==2.1
urllib3==1.26.2

mysql-connector-python~=8.0.22